requirements.txt
Django==1.8.2
Markdown==2.6.2
PyYAML==3.11
Werkzeug==0.10.4
django-braces==1.8.0
django-cors-headers==1.1.0
django-debug-toolbar==1.3.0
django-extensions==1.5.5
django-filter==0.10.0
djangorestframework==3.1.2
gnureadline==6.3.3
ipdb==0.8
ipython==3.1.0
six==1.9.0
sqlparse==0.1.15
wsgiref==0.1.2
